First version of integration of IMU-DVL-Depth meter.
Still to add :
- Depth offset service
- Heading offset service
- Position offset service

Everything must be test with the submarine.

There are functions (quat to euler, etc.) to test in Atlas

NavigationNode now loops and check if the three provider have update their position. If they had, update and push new message.

*Data.cc and .h are the message receiver from the device and handle updating their internal values.

NavigationDevice is the base class for those devices. For now it is only update functions.