Adding the reset functionnality and test working with device simulator.
Adding the device simulator, which emulate expected behavior from IMU, DVL and depth meter.
It use our new F310 keypad and works well.

Adding the TF script for enu to ned. Before it was launched in the launch file, with bash command, now in bash script to mimic the tf_ned_to_body.

This commit is all tested in rviz. It works there, so if there is issue in the test, it is more about what data does the DVL send...